// Frontend/src/components/DestinosCard/DataDestination.interface.ts
export interface DataDestination{
    id: number;
    images: string[];
    city: string;
    title: string;
    subtitle: string;
    location: string;
    locationTitle: string;
    highSeason: string;
    highSeasonTitle: string;
    celebrations: string;
    celebrationsTitle: string;
    phase: string;
    tag: string[];
}